Refactor: LogAction 이동 및 수정
- LogAction 지문 도메인 내부로 이동
- LogAction 에 편의점 및 운동 추가
