+++ src/main/java/com/example/fingerprint_backend/domain/fingerprint/types/LogAction.java
@@ -1,0 +1,5 @@
+package com.example.fingerprint_backend.domain.fingerprint.types;
+
+public enum LogAction {
+    등교, 하교, 식사, 도서관, 기타, 복귀, 편의점, 운동
+}
