+++ src/main/java/com/example/fingerprint_backend/dto/CreateLogDto.java
@@ -1,6 +1,6 @@
 package com.example.fingerprint_backend.dto;
 
-import com.example.fingerprint_backend.types.LogAction;
+import com.example.fingerprint_backend.domain.fingerprint.types.LogAction;
 import lombok.Getter;
 
 @Getter
