--- src/main/java/com/example/fingerprint_backend/types/LogAction.java
@@ -1,5 +1,0 @@
-package com.example.fingerprint_backend.types;
-
-public enum LogAction {
-    등교, 하교, 식사, 도서관, 기타, 복귀
-}
