+++ src/test/java/com/example/fingerprint_backend/domain/fingerprint/service/cycle/CycleCreateTest.java
@@ -12,7 +12,7 @@ import com.example.fingerprint_backend.entity.MemberEntity;
 import com.example.fingerprint_backend.entity.SchoolClass;
 import com.example.fingerprint_backend.service.AuthService;
 import com.example.fingerprint_backend.service.CleanManagementService;
-import com.example.fingerprint_backend.types.LogAction;
+import com.example.fingerprint_backend.domain.fingerprint.types.LogAction;
 import com.example.fingerprint_backend.types.MemberRole;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
